selecting the space with the empty room before attempt to interact with it
