+++ vector/src/androidTest/java/im/vector/app/ui/robot/space/SpaceRobot.kt
@@ -16,9 +16,17 @@
 
 package im.vector.app.ui.robot.space
 
+import androidx.recyclerview.widget.RecyclerView
+import androidx.test.espresso.Espresso
+import androidx.test.espresso.contrib.RecyclerViewActions
+import androidx.test.espresso.matcher.ViewMatchers
 import com.adevinta.android.barista.interaction.BaristaClickInteractions.clickOn
 import com.adevinta.android.barista.interaction.BaristaDrawerInteractions.openDrawer
+import com.adevinta.android.barista.internal.viewaction.ClickChildAction
 import im.vector.app.R
+import im.vector.app.espresso.tools.waitUntilDialogVisible
+import im.vector.app.espresso.tools.waitUntilViewVisible
+import org.hamcrest.Matchers
 
 class SpaceRobot {
 
@@ -34,5 +42,23 @@ class SpaceRobot {
             openMenu(spaceName)
             block()
         }
+    }
+
+    fun openMenu(spaceName: String) {
+        waitUntilViewVisible(ViewMatchers.withId(R.id.groupListView))
+        Espresso.onView(ViewMatchers.withId(R.id.groupListView))
+                .perform(
+                        RecyclerViewActions.actionOnItem<RecyclerView.ViewHolder>(
+                                ViewMatchers.hasDescendant(Matchers.allOf(ViewMatchers.withId(R.id.groupNameView), ViewMatchers.withText(spaceName))),
+                                ClickChildAction.clickChildWithId(R.id.groupTmpLeave)
+                        ).atPosition(0)
+                )
+        waitUntilDialogVisible(ViewMatchers.withId(R.id.spaceNameView))
+    }
+
+    fun selectSpace(spaceName: String) {
+        openDrawer()
+        waitUntilViewVisible(ViewMatchers.withId(R.id.groupListView))
+        clickOn(spaceName)
     }
 }
